Fix duplicate query log processing
Query logs were not being correctly added when there were duplicates.
Now the queries are all put into a Guava MultiMap and then put into the
Trie. Before, it was putting the contents of each document into the
trie. However, this meant that if there were any duplicate keys (the
query strings themselves), they would wipe out the existing entries
entirely.

This is now fixed. The number of entries in the Trie now match the
total number of lines in each of the five AOL query log files.